Move image helpers to its own package
