+++ image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/ImageLoader.scala
@@ -3,7 +3,6 @@ package eu.joaocosta.minart.graphics.image
 import java.io.InputStream
 
 import eu.joaocosta.minart.graphics._
-import eu.joaocosta.minart.graphics.image.Helpers._
 
 trait ImageLoader {
   def loadImage(is: InputStream): Either[String, RamSurface]
